personal-rag/requirements.txt
streamlit==1.39.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
PyMuPDF==1.24.9
pytesseract==0.3.10
Pillow==10.4.0
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
beautifulsoup4==4.12.3
ebooklib==0.18
requests==2.32.3
pytest==8.3.2
numpy==1.26.4
